+++ HISTORY.rst
@@ -4,7 +4,8 @@ History
 0.2.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Introducing subset fields that can be used to query results that are collected
+  only for subsets of the model, like the 2D section.
 
 
 0.2.3 (2018-05-14)